refactor: consolidate constants into centralized constants directory
